Add files via upload
The data analysis has been carried out based on the txt file that is exported from the WhatsApp group chats. The total number of messages, multimedia, emojis and links has been calculated, as well as the graph of the behavior of the number of messages over time. In relation to the statistics per member, the number of member words per message and number of messages, multimedia, links and emojis for each one were obtained. Also, the most used emojis were determined and a pie chart was made for a better observation. Likewise, it was possible to know who the most active members are and the days and hours that the group is most active. Finally, the word cloud of the most used words in the chat was graphed.